Mintified some more files. Tiny HACKING change.

+++ src/m_random.h
@@ -1,4 +1,4 @@
-// Emacs style mode select   -*- C++ -*- 
+// Emacs style mode select   -*- C++ -*-
 //-----------------------------------------------------------------------------
 //
 // Copyright(C) 1993-1996 Id Software, Inc.
@@ -21,27 +21,22 @@
 //
 // DESCRIPTION:
 //
-//    
+//
 //-----------------------------------------------------------------------------
-
 
 #ifndef __M_RANDOM__
 #define __M_RANDOM__
 
-
 #include "doomtype.h"
-
-
 
 // Returns a number from 0 to 255,
 // from a lookup table.
-int M_Random (void);
+uint8_t M_Random(void);
 
 // As M_Random, but used only by the play simulation.
-int P_Random (void);
+uint8_t P_Random(void);
 
 // Fix randoms for demos.
-void M_ClearRandom (void);
-
+void M_ClearRandom(void);
 
 #endif
